Fix QueryParsingErrorsSuite
