+++ sql/core/src/test/scala/org/apache/spark/sql/errors/QueryParsingErrorsSuite.scala
@@ -18,24 +18,10 @@
 package org.apache.spark.sql.errors
 
 import org.apache.spark.sql.QueryTest
-import org.apache.spark.sql.catalyst.parser.ParseException
-import org.apache.spark.sql.test.SharedSparkSession
 
 // Turn of the length check because most of the tests check entire error messages
 // scalastyle:off line.size.limit
-class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
-  def validateParsingError(
-      sqlText: String,
-      errorClass: String,
-      sqlState: String,
-      message: String): Unit = {
-    val e = intercept[ParseException] {
-      sql(sqlText)
-    }
-    assert(e.getErrorClass === errorClass)
-    assert(e.getSqlState === sqlState)
-    assert(e.getMessage === message)
-  }
+class QueryParsingErrorsSuite extends QueryTest with QueryErrorsSuiteBase {
 
   test("UNSUPPORTED_FEATURE: LATERAL join with NATURAL join not supported") {
     validateParsingError(
@@ -43,8 +29,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "UNSUPPORTED_FEATURE",
       sqlState = "0A000",
       message =
-        """
-          |The feature is not supported: "LATERAL" join with "NATURAL" join.(line 1, pos 14)
+        """The feature is not supported: "LATERAL" join with "NATURAL" join.(line 1, pos 14)
           |
           |== SQL ==
           |SELECT * FROM t1 NATURAL JOIN LATERAL (SELECT c1 + c2 AS c2)
@@ -58,8 +43,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "UNSUPPORTED_FEATURE",
       sqlState = "0A000",
       message =
-        """
-          |The feature is not supported: "LATERAL" join with "USING" join.(line 1, pos 14)
+        """The feature is not supported: "LATERAL" join with "USING" join.(line 1, pos 14)
           |
           |== SQL ==
           |SELECT * FROM t1 JOIN LATERAL (SELECT c1 + c2 AS c2) USING (c2)
@@ -74,8 +58,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
         errorClass = "UNSUPPORTED_FEATURE",
         sqlState = "0A000",
         message =
-          s"""
-            |The feature is not supported: "LATERAL" join type "$joinType".(line 1, pos 14)
+          s"""The feature is not supported: "LATERAL" join type "$joinType".(line 1, pos 14)
             |
             |== SQL ==
             |SELECT * FROM t1 $joinType JOIN LATERAL (SELECT c1 + c2 AS c3) ON c2 = c3
@@ -98,8 +81,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
         errorClass = "INVALID_SQL_SYNTAX",
         sqlState = "42000",
         message =
-          s"""
-            |Invalid SQL syntax: "LATERAL" can only be used with subquery.(line 1, pos $pos)
+          s"""Invalid SQL syntax: "LATERAL" can only be used with subquery.(line 1, pos $pos)
             |
             |== SQL ==
             |$sqlText
@@ -114,8 +96,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "UNSUPPORTED_FEATURE",
       sqlState = "0A000",
       message =
-        """
-          |The feature is not supported: "NATURAL CROSS JOIN".(line 1, pos 14)
+        """The feature is not supported: "NATURAL CROSS JOIN".(line 1, pos 14)
           |
           |== SQL ==
           |SELECT * FROM a NATURAL CROSS JOIN b
@@ -129,8 +110,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        """
-          |Invalid SQL syntax: The definition of window `win` is repetitive.(line 1, pos 31)
+        """Invalid SQL syntax: The definition of window `win` is repetitive.(line 1, pos 31)
           |
           |== SQL ==
           |SELECT min(a) OVER win FROM t1 WINDOW win AS win, win AS win2
@@ -144,8 +124,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        """
-          |Invalid SQL syntax: Window reference `win` is not a window specification.(line 1, pos 31)
+        """Invalid SQL syntax: Window reference `win` is not a window specification.(line 1, pos 31)
           |
           |== SQL ==
           |SELECT min(a) OVER win FROM t1 WINDOW win AS win
@@ -159,8 +138,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        """
-          |Invalid SQL syntax: Cannot resolve window reference `win2`.(line 1, pos 31)
+        """Invalid SQL syntax: Cannot resolve window reference `win2`.(line 1, pos 31)
           |
           |== SQL ==
           |SELECT min(a) OVER win FROM t1 WINDOW win AS win2
@@ -174,8 +152,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "UNSUPPORTED_FEATURE",
       sqlState = "0A000",
       message =
-        """
-          |The feature is not supported: "TRANSFORM" does not support "DISTINCT"/"ALL" in inputs(line 1, pos 17)
+        """The feature is not supported: "TRANSFORM" does not support "DISTINCT"/"ALL" in inputs(line 1, pos 17)
           |
           |== SQL ==
           |SELECT TRANSFORM(DISTINCT a) USING 'a' FROM t
@@ -190,8 +167,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "UNSUPPORTED_FEATURE",
       sqlState = "0A000",
       message =
-        """
-          |The feature is not supported: "TRANSFORM" with serde is only supported in hive mode(line 1, pos 0)
+        """The feature is not supported: "TRANSFORM" with serde is only supported in hive mode(line 1, pos 0)
           |
           |== SQL ==
           |SELECT TRANSFORM(a) ROW FORMAT SERDE 'org.apache.hadoop.hive.serde2.OpenCSVSerde' USING 'a' FROM t
@@ -205,8 +181,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        """
-          |Invalid SQL syntax: Too many arguments for transform `years`(line 1, pos 44)
+        """Invalid SQL syntax: Too many arguments for transform `years`(line 1, pos 44)
           |
           |== SQL ==
           |CREATE TABLE table(col int) PARTITIONED BY (years(col,col))
@@ -220,8 +195,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        """
-          |Invalid SQL syntax: Unsupported function name `ns`.`db`.`func`(line 1, pos 14)
+        """Invalid SQL syntax: Unsupported function name `ns`.`db`.`func`(line 1, pos 14)
           |
           |== SQL ==
           |SELECT * FROM ns.db.func()
@@ -235,8 +209,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        """
-          |Invalid SQL syntax: "SHOW" `sys` "FUNCTIONS" not supported(line 1, pos 5)
+        """Invalid SQL syntax: "SHOW" `sys` "FUNCTIONS" not supported(line 1, pos 5)
           |
           |== SQL ==
           |SHOW sys FUNCTIONS
@@ -250,8 +223,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        s"""
-          |Invalid SQL syntax: Invalid pattern in "SHOW FUNCTIONS": `f1`. It must be a STRING literal.(line 1, pos 21)
+        """Invalid SQL syntax: Invalid pattern in "SHOW FUNCTIONS": `f1`. It must be a STRING literal.(line 1, pos 21)
           |
           |== SQL ==
           |SHOW FUNCTIONS IN db f1
@@ -262,8 +234,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        s"""
-          |Invalid SQL syntax: Invalid pattern in "SHOW FUNCTIONS": `f1`. It must be a STRING literal.(line 1, pos 26)
+        """Invalid SQL syntax: Invalid pattern in "SHOW FUNCTIONS": `f1`. It must be a STRING literal.(line 1, pos 26)
           |
           |== SQL ==
           |SHOW FUNCTIONS IN db LIKE f1
@@ -286,8 +257,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        s"""
-          |Invalid SQL syntax: $errorDesc
+        s"""Invalid SQL syntax: $errorDesc
           |
           |== SQL ==
           |
@@ -313,8 +283,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        s"""
-          |Invalid SQL syntax: $errorDesc
+        s"""Invalid SQL syntax: $errorDesc
           |
           |== SQL ==
           |
@@ -338,8 +307,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        """
-          |Invalid SQL syntax: Unsupported function name `ns`.`db`.`func`(line 2, pos 0)
+        """Invalid SQL syntax: Unsupported function name `ns`.`db`.`func`(line 2, pos 0)
           |
           |== SQL ==
           |
@@ -365,8 +333,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        s"""
-          |Invalid SQL syntax: $errorDesc
+        s"""Invalid SQL syntax: $errorDesc
           |
           |== SQL ==
           |
@@ -386,8 +353,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "INVALID_SQL_SYNTAX",
       sqlState = "42000",
       message =
-        s"""
-          |Invalid SQL syntax: $errorDesc
+        s"""Invalid SQL syntax: $errorDesc
           |
           |== SQL ==
           |DROP TEMPORARY FUNCTION db.func
@@ -401,8 +367,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "DUPLICATE_KEY",
       sqlState = "23000",
       message =
-        """
-          |Found duplicate keys `p1`(line 1, pos 29)
+        """Found duplicate keys `p1`(line 1, pos 29)
           |
           |== SQL ==
           |INSERT OVERWRITE TABLE table PARTITION(p1='1', p1='1') SELECT 'col1', 'col2'
@@ -416,8 +381,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "DUPLICATE_KEY",
       sqlState = "23000",
       message =
-        """
-          |Found duplicate keys `key1`(line 1, pos 39)
+        """Found duplicate keys `key1`(line 1, pos 39)
           |
           |== SQL ==
           |ALTER TABLE dbx.tab1 SET TBLPROPERTIES ('key1' = '1', 'key1' = '2')
@@ -431,8 +395,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_EMPTY_STATEMENT",
       sqlState = "42000",
       message =
-        """
-          |Syntax error, unexpected empty statement(line 1, pos 0)
+        """Syntax error, unexpected empty statement(line 1, pos 0)
           |
           |== SQL ==
           |
@@ -444,8 +407,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_EMPTY_STATEMENT",
       sqlState = "42000",
       message =
-        s"""
-           |Syntax error, unexpected empty statement(line 1, pos 3)
+        s"""Syntax error, unexpected empty statement(line 1, pos 3)
            |
            |== SQL ==
            |${"   "}
@@ -457,8 +419,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_EMPTY_STATEMENT",
       sqlState = "42000",
       message =
-        s"""
-           |Syntax error, unexpected empty statement(line 2, pos 0)
+        s"""Syntax error, unexpected empty statement(line 2, pos 0)
            |
            |== SQL ==
            |${" "}
@@ -473,8 +434,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_SYNTAX_ERROR",
       sqlState = "42000",
       message =
-        s"""
-          |Syntax error at or near end of input(line 1, pos 16)
+        s"""Syntax error at or near end of input(line 1, pos 16)
           |
           |== SQL ==
           |$sqlText
@@ -488,8 +448,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_SYNTAX_ERROR",
       sqlState = "42000",
       message =
-        s"""
-          |Syntax error at or near '1': extra input '1'(line 1, pos 9)
+        """Syntax error at or near '1': extra input '1'(line 1, pos 9)
           |
           |== SQL ==
           |select 1 1
@@ -501,8 +460,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_SYNTAX_ERROR",
       sqlState = "42000",
       message =
-        s"""
-          |Syntax error at or near 't': extra input 't'(line 2, pos 12)
+        """Syntax error at or near 't': extra input 't'(line 2, pos 12)
           |
           |== SQL ==
           |select *
@@ -517,8 +475,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_SYNTAX_ERROR",
       sqlState = "42000",
       message =
-        """
-          |Syntax error at or near 'from'(line 1, pos 27)
+        """Syntax error at or near 'from'(line 1, pos 27)
           |
           |== SQL ==
           |select * from r order by q from t
@@ -530,8 +487,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_SYNTAX_ERROR",
       sqlState = "42000",
       message =
-        """
-          |Syntax error at or near 'from'(line 4, pos 0)
+        """Syntax error at or near 'from'(line 4, pos 0)
           |
           |== SQL ==
           |select *
@@ -549,8 +505,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_SYNTAX_ERROR",
       sqlState = "42000",
       message =
-        """
-          |Syntax error at or near end of input(line 1, pos 14)
+        """Syntax error at or near end of input(line 1, pos 14)
           |
           |== SQL ==
           |select count(*
@@ -562,8 +517,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_SYNTAX_ERROR",
       sqlState = "42000",
       message =
-        """
-          |Syntax error at or near end of input(line 1, pos 18)
+        """Syntax error at or near end of input(line 1, pos 18)
           |
           |== SQL ==
           |select 1 as a from
@@ -578,8 +532,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_SYNTAX_ERROR",
       sqlState = "42000",
       message =
-        """
-          |Syntax error at or near 'join_': missing 'JOIN'(line 1, pos 21)
+        """Syntax error at or near 'join_': missing 'JOIN'(line 1, pos 21)
           |
           |== SQL ==
           |select * from a left join_ b on a.id = b.id
@@ -591,8 +544,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_SYNTAX_ERROR",
       sqlState = "42000",
       message =
-        """
-          |Syntax error at or near 'is'(line 1, pos 32)
+        """Syntax error at or near 'is'(line 1, pos 32)
           |
           |== SQL ==
           |select * from test where test.t is like 'test'
@@ -604,8 +556,7 @@ class QueryParsingErrorsSuite extends QueryTest with SharedSparkSession {
       errorClass = "PARSE_SYNTAX_ERROR",
       sqlState = "42000",
       message =
-        """
-          |Syntax error at or near 'NOT'(line 1, pos 27)
+        """Syntax error at or near 'NOT'(line 1, pos 27)
           |
           |== SQL ==
           |SELECT * FROM test WHERE x NOT NULL
